+++ cell_path.go
@@ -1,6 +1,7 @@
 package nu
 
 import (
+	"errors"
 	"fmt"
 	"math"
 
@@ -35,6 +36,9 @@ type PathMember interface {
 	// accessed - the path access will just return Nothing instead.
 	Optional() bool
 
+	// Is the path element case-sensitive (true) or case-insensitive (false).
+	CaseSensitive() bool
+
 	Span() Span
 
 	string() string
@@ -65,19 +69,19 @@ func (cp CellPath) Length() int {
 }
 
 func (cp *CellPath) AddInteger(value uint, optional bool) {
-	cp.Members = append(cp.Members, pathItem[uint]{value: value, optional: optional})
+	cp.Members = append(cp.Members, pathItem[uint]{value: value, optional: optional, casing: true})
 }
 
 func (cp *CellPath) AddIntegerSpan(value uint, optional bool, span Span) {
-	cp.Members = append(cp.Members, pathItem[uint]{value: value, optional: optional, span: span})
+	cp.Members = append(cp.Members, pathItem[uint]{value: value, optional: optional, casing: true, span: span})
 }
 
-func (cp *CellPath) AddString(value string, optional bool) {
-	cp.Members = append(cp.Members, pathItem[string]{value: value, optional: optional})
+func (cp *CellPath) AddString(value string, optional, caseSensitive bool) {
+	cp.Members = append(cp.Members, pathItem[string]{value: value, optional: optional, casing: caseSensitive})
 }
 
-func (cp *CellPath) AddStringSpan(value string, optional bool, span Span) {
-	cp.Members = append(cp.Members, pathItem[string]{value: value, optional: optional, span: span})
+func (cp *CellPath) AddStringSpan(value string, optional, caseSensitive bool, span Span) {
+	cp.Members = append(cp.Members, pathItem[string]{value: value, optional: optional, casing: caseSensitive, span: span})
 }
 
 func (cp *CellPath) decodeMsgpack(dec *msgpack.Decoder, p *Plugin) error {
@@ -123,6 +127,7 @@ func (cp *CellPath) encodeMsgpack(enc *msgpack.Encoder, p *Plugin) error {
 type pathItem[T uint | string] struct {
 	value    T
 	optional bool
+	casing   bool // is the (string) member case sensitive?
 	span     Span
 }
 
@@ -130,6 +135,9 @@ func (pi pathItem[T]) string() string {
 	opt := ""
 	if pi.optional {
 		opt = "?"
+	}
+	if !pi.casing && pi.Type() == PathVariantString {
+		opt += "!"
 	}
 	return fmt.Sprintf("%v%s", pi.value, opt)
 }
@@ -157,6 +165,8 @@ func (pi pathItem[T]) PathStr() string {
 
 func (pi pathItem[T]) Optional() bool { return pi.optional }
 
+func (pi pathItem[T]) CaseSensitive() bool { return pi.casing }
+
 func (pi pathItem[T]) Span() Span { return pi.span }
 
 func (pi pathItem[T]) encodeMsgpack(enc *msgpack.Encoder, p *Plugin) error {
@@ -167,7 +177,7 @@ func (pi pathItem[T]) encodeMsgpack(enc *msgpack.Encoder, p *Plugin) error {
 	if err := encodeMapStart(enc, vtyp); err != nil {
 		return err
 	}
-	if err := enc.EncodeMapLen(3); err != nil {
+	if err := enc.EncodeMapLen(4); err != nil {
 		return err
 	}
 	if err := enc.EncodeString("val"); err != nil {
@@ -182,6 +192,13 @@ func (pi pathItem[T]) encodeMsgpack(enc *msgpack.Encoder, p *Plugin) error {
 	if err := pi.span.encodeMsgpack(enc); err != nil {
 		return fmt.Errorf("encoding span: %w", err)
 	}
+	s := "Sensitive"
+	if !pi.casing {
+		s = "Insensitive"
+	}
+	if err := encodeString(enc, "casing", s); err != nil {
+		return err
+	}
 	return encodeBoolean(enc, "optional", pi.optional)
 }
 
@@ -195,7 +212,7 @@ func decodePathMember(dec *msgpack.Decoder, p *Plugin) (PathMember, error) {
 		return nil, err
 	}
 
-	sVal, iVal, span, opt := "", uint(0), Span{}, false
+	sVal, iVal, span, opt, casing := "", uint(0), Span{}, false, true
 	for idx := range cnt {
 		key, err := dec.DecodeString()
 		if err != nil {
@@ -215,6 +232,20 @@ func decodePathMember(dec *msgpack.Decoder, p *Plugin) (PathMember, error) {
 			err = span.decodeMsgpack(dec)
 		case "optional":
 			opt, err = dec.DecodeBool()
+		case "casing":
+			var s string
+			if s, err = dec.DecodeString(); err == nil {
+				switch s {
+				case "Sensitive":
+					casing = true
+				case "Insensitive":
+					casing = false
+				default:
+					err = fmt.Errorf("unsupported value %q", s)
+				}
+			}
+		default:
+			err = errors.New("unsupported key")
 		}
 		if err != nil {
 			return nil, fmt.Errorf("decoding key %q: %w", key, err)
@@ -223,9 +254,9 @@ func decodePathMember(dec *msgpack.Decoder, p *Plugin) (PathMember, error) {
 
 	switch itemType {
 	case "Int":
-		return pathItem[uint]{value: iVal, span: span, optional: opt}, nil
+		return pathItem[uint]{value: iVal, span: span, optional: opt, casing: casing}, nil
 	case "String":
-		return pathItem[string]{value: sVal, span: span, optional: opt}, nil
+		return pathItem[string]{value: sVal, span: span, optional: opt, casing: casing}, nil
 	}
 	return nil, fmt.Errorf("unsupported CellPath member type %s", itemType)
 }